+++ src/main/scala/net/fehmicansaglam/tepkin/protocol/result/Result.scala
@@ -4,7 +4,7 @@ import net.fehmicansaglam.tepkin.bson.BsonDocument
 
 protected[result] trait Result
 
-case class CountResult(missing: Option[Boolean] = None, n: Long, ok: Boolean) extends Result
+case class CountResult(missing: Option[Boolean] = None, n: Int, ok: Boolean) extends Result
 
 case class DeleteResult(n: Option[Int], code: Option[Int], errmsg: Option[String], ok: Boolean) extends Result
